[DPE-6603] Add extra checks for `DeploymentDescription` existence (#548)
Currently, if the cluster is not yet ready and the description not yet
available, set password action throws an exception. This PR relates to:
#547.

It adds checks for the deployment description on actions and some
key-points in the base charm. This PR reviews the base charm calls to
the `.deployment_desc()` function and ensures we have appropriate checks
each time.

---------

Co-authored-by: Mehdi Bendriss <bendrissmehdi@gmail.com>
Co-authored-by: smethnani <82812166+smethnani@users.noreply.github.com>