[DPE-6582]  App not inheriting cluster name in large deployments (#552)
## Issue #549

When relating two apps in a large deployment, there is a check which
blocks the relation if the cluster names don't match despite
inherit-cluster-name being in the pending directives.

## Solution
Add a check for inherit-cluster-name in the pending directives before
blocking the relation. Additionally, store whether or not the cluster
name was auto generated and only inherit the cluster name if it was auto
generated.