[DPE-6765] Avoid restarting if the node is not up && reduce log errors in low-level methods (#571)
This PR does two things:
1) adds a check to avoid unnecessary restarts whenever the unit is
already powered down
2) Reduces the log level of too low-level methods

For the item (2): this is a normal deployment:
https://pastebin.ubuntu.com/p/fwZvJyzW87/

We can see there are two ERROR entries:
```
unit-opensearch-0: 10:49:27 ERROR unit.opensearch/0.juju-log certificates:3: err:  / out: keytool error: java.lang.Exception: Keystore file does not exist: /var/snap/opensearch/current/etc/opensearch/certificates/ca.p12
```

Coming from
[here](https://github.com/canonical/opensearch-operator/blob/2/edge/lib/charms/opensearch/v0/opensearch_tls.py#L548).
In this case, this ERROR is not really an error, as the check shown
previously will not raise again. That failing command is just used to
move a potential CA to `old-ca` alias or be ignored.

Therefore, that ERROR will essentially mean nothing on a
troubleshooting.

And:
```
unit-opensearch-0: 10:49:33 ERROR unit.opensearch/0.juju-log Cannot connect to the OpenSearch server...
```

This is raised by `is_started`. The `is_started` method is used mostly
internally within `opensearch_distro` as a check if we really need to
start/stop/restart. In these cases, there is no reason to have that as
an ERROR.

Closes #570