[DPE-7566] - Separate data and error fields in peer-cluster relation data (#653)
## Issue
The peer-cluster-orchestrator relation data currently sets either an
error or data. This causes some issues when applications need to access
information in the data field but the orchestrator has set the error
field.

## Solution
Separate the error and data creation and send both, unless the relation
should be severed, in which case it should only send the error data.

Addresses [DPE-7566]

[DPE-7566]:
https://warthogs.atlassian.net/browse/DPE-7566?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ