[DPE-7076] Data node fails to start (#671)
## Issue
If the presence of the data node is detected during the first call to
`refresh_relation_data`, a `'Waiting for every unit in
<main-orchestrator> to start'` error state is set in the peer-cluster
relation data, blocking the data node from starting and from
initializing the security index, which in turn blocks the units from
other applications from starting and results in a deadlock.

## Solution
Only set the error if the main orchestrator detects a data node in the
cluster and the security index has been initialized.

This PR also re-ables the remove-orchestrators tests and clears the
`Requesting lock on operation start` status in post-start-init for the
cases where opensearch takes longer than expected to start and
`is_started` is detected at a later event

Fixes #631 / [DPE-7076] and [DPE-7527]


[DPE-7076]:
https://warthogs.atlassian.net/browse/DPE-7076?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ
[DPE-7527]:
https://warthogs.atlassian.net/browse/DPE-7527?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ