[DPE-7880] Wrap credential assignments for backwards compatibility (#683)
The credentials in the peer cluster relation data were not passed as
secrets in older revisions, breaking upgrades from those revisions in
large deployments.

This PR adds a wrapper to resolve the credentials as secrets only if
they are uri's using the `secret://` scheme.

The value in `credentials["admin_tls"]` was previously a dictionary, not
a string. The case is handled directly in `rel_data_from_str` instead of
using the wrapper.

Addresses [DPE-7880]


[DPE-7880]:
https://warthogs.atlassian.net/browse/DPE-7880?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ