[DPE-6808] Fix channel name for 'data-integrator' in README.md (#692)
## Description of issue or feature:
This PR updates the channel name for 'data-integrator' charm in the
README.md. In the README.md we show how to integrate the opensearch
charm with the data-integrator charm. However the charm channel is not
valid as reported in #590.
fix #590 

## Solution:
Updated channel name to 'stable' following the docs
[here](https://charmhub.io/opensearch/docs/t-integrate?channel=2/edge).

I also made some quick fixes that I think will improve the quality and
clarity of the docs.


## How was this change tested?
- [ ] Manually
- [ ] Unit tests
- [ ] Integration tests




## Checklist
- [X] I have updated README.md