[DPE-8017] Add jwt authentication (#696)
## Description of issue or feature:
This PR adds JSON Web Token (JWT) authentication to Opensearch.

## Solution:
The configuration of JWT authentication is provided by the
`jwt-integrator` charm via a new relation called `jwt-configuration`.
The configuration parameters are shared via relation data and relation
secrets.

The business logic and event handling is bundled in a new class called
`JwtHandler`. On the relevant events, the logic applies the
configuration to the Opensearch security configuration file and updates
the security index by running the `securityadmin` script.

In large deployments, the relation must be configured to the
Main-Orchestrator app. If not, the events will be ignored and a blocked
status will be displayed.

## How was this change tested?
- [X] Manually
- [ ] Unit tests
- [X] Integration tests

Manual tests have been executed with both small and large deployments in
remote-VM environments provisioned with testflinger.

The provided integration test in `test_jwt_auth.py` covers:
- small deployment
- large deployment with data-only nodes
- invalid relation: blocked status and no access with the configured JWT
  - valid relation: data can be accessed with the configured JWT

## Checklist
- [X] I have added or updated any relevant documentation -> will happen
in a Discourse post once this PR is merged.
- [X] I have cleaned any remaining cloud resources from my accounts.