fix([DPE-8408]): Isolate OpenSearchNotFullyReadyError and add retry machanisme (#704)
## Description of issue or feature:
The charm is blocked on deployment for slow cluster startups

## Solution:
Isolate OpenSearchNotFullyReadyError and add retry machanisme. Re-open
the issue for the cluster ID mismatch.
* Added a retry loop using `tenacity.Retrying` in `_post_start_init` to
repeatedly check if the node is up, with exponential backoff, before
proceeding. This ensures the node is fully ready and avoids premature
readiness signaling.
[[1]](diffhunk://#diff-5c0d89f838ffd5a9ecf2288b3698320fda2d2b2b737361b6e10adb55d1929f43R117)
[[2]](diffhunk://#diff-5c0d89f838ffd5a9ecf2288b3698320fda2d2b2b737361b6e10adb55d1929f43R1214-R1221)
* Refined the exception handling in `_start_opensearch` to catch
`OpenSearchStartError` and `OpenSearchUserMgmtError` together, log
warnings for these errors, and set the unit status to `BlockedStatus`
with `ServiceStartError`. Logging for `OpenSearchNotFullyReadyError` was
moved to a debug message, clarifying the node readiness state.
[[1]](diffhunk://#diff-5c0d89f838ffd5a9ecf2288b3698320fda2d2b2b737361b6e10adb55d1929f43L1169-R1174)
[[2]](diffhunk://#diff-5c0d89f838ffd5a9ecf2288b3698320fda2d2b2b737361b6e10adb55d1929f43L1180-R1185)

## How was this change tested?
- [x] Manually
- [x] Unit tests
- [x] Integration tests

---------

Co-authored-by: Gulsum Atici <gulsum.atici@canonical.com>