Defer departing event if service is down or leader is departing (#715)
## Description of issue or feature:
<!--- Please describe in detail what this PR is about. And a reference
link to a Github issue or Jira ticket. -->
Removing the main leader unit results in incorrect information in
cluster_fleet_apps as departed events for all units in subclusters are
processed on the main leader before it shuts down.

Additionally, nodes-config will be incorrect if a peer unit departs
while the OpenSearch service is down on the leader unit.
## Solution:
<!--- Please describe in detail what the solution implemented in this PR
is. The changes made etc. -->
Defer the departed event if the node is not up. If the service returns
and the leader unit has not departed, cluster_fleet_apps and/or
nodes_config will be corrected once the event is re-emitted.
## How was this change tested?
- [x] Manually
- [ ] Unit tests
- [ ] Integration tests

<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

Closes #709/[DPE-8499] 


[DPE-8499]:
https://warthogs.atlassian.net/browse/DPE-8499?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ