Fixing failing unit tests on TiCs workflows (#731)
## Description of issue or feature:
This PR addresses issue #716 . The tics workflows was failing due to
unit tests that were not passing on self-hosted runners
[[1]](https://github.com/canonical/opensearch-operator/actions/workflows/tics_run_sh_ghaction_test.yml).
After an investigation of the failing unit tests, we found that the
function`is_reachable` in the`helpers_networking`[ [2]
](https://github.com/canonical/opensearch-operator/blob/387544298dfd447c2ae8a62bc811365a6f673c2d/lib/charms/opensearch/v0/helper_networking.py#L87C48-L88C1)is
giving some weird false-positive results. This means that it was giving
a success result when using IPs that don't exist or don't work when
creating a socket connection to. After tracking this issue, we found
that the issue was due to the tight configuration of self-hosted
runners, they have an aproxy configuration that allow only a certain
range of IPs.

## Solution:
The Self-Hosted runners team proposed to add the unit test IP to the
excluded IPs using this command `sudo nft add element ip aproxy exclude
{ 1.1.1.1 }`. As a result we have added this workaround to the github
actions workflow.

# Note:
I have temporary updated the workflow to run on pull request making sure
to validate the PR and then remove the trigger.

## How was this change tested?
- [ ] Manually
- [X] Unit tests
- [ ] Integration tests

[1]
https://github.com/canonical/opensearch-operator/actions/workflows/tics_run_sh_ghaction_test.yml
[2]
https://github.com/canonical/opensearch-operator/blob/387544298dfd447c2ae8a62bc811365a6f673c2d/lib/charms/opensearch/v0/helper_networking.py#L87C48-L88C1