[DPE-8405] Restrict relation with OAuth provider to main-orchestrator (#730)
## Description of issue or feature:
Fixes https://github.com/canonical/opensearch-operator/issues/699.

## Solution:
This PR restricts the OAuth relation in large deployments of Opensearch
to only be created with the main-orchestrator application of the
cluster. Otherwise the OAuth configuration will not be applied to
Opensearch and a `blocked` status will be displayed.

The solution is in line with the one for JWT authentication
(implementation and UX are the same).

## How was this change tested?
- [X] Manually
- [ ] Unit tests
- [X] Integration tests

The PR also adds integration test coverage for this use case:
- creating a large deployment consisting of cluster-manager-only and
data-only nodes
- verifying `blocked` status and OAuth access not possible if relation
is created with `Data` application
- when the relation is established with the `Main` application, OAuth
access is possible on the `Data` nodes.

Furthermore, manual tests for both small and large deployments have been
conducted, ensuring that OAuth and JWT authentication can be used in
parallel and both work.