fix([DPE-7059]): Raise on the last cluster manager unit (#733)
## Description of issue or feature:
<!--- Please describe in detail what this PR is about. And a reference
link to a Github issue or Jira ticket. -->

This pull request introduces a new error handling mechanism for the
scenario where no cluster manager units are available in an OpenSearch
cluster. The changes ensure that the system will raise a specific
exception and log an error if all cluster managers are lost, helping to
prevent cluster state loss and improve operational robustness.

## Solution:
<!--- Please describe in detail what the solution implemented in this PR
is. The changes made etc. -->

**Error handling improvements:**

* Added a new exception class `OpenSearchNoClusterManagersError` to
`opensearch_exceptions.py` for cases when there are no cluster managers
in the cluster.
* Updated the imports in `opensearch_base_charm.py` to include
`OpenSearchNoClusterManagersError`.
* In the `_on_opensearch_data_storage_detaching` method, added logic to
check for the presence of cluster managers and raise
`OpenSearchNoClusterManagersError` if none are up, with an appropriate
error log message.

**Cluster manager status checks:**

* Added the method `does_cluster_have_cm_up` to
`opensearch_peer_clusters.py` to determine if at least one cluster
manager unit is up in the fleet.


## How was this change tested?
- [x] Manually
- [ ] Unit tests
- [ ] Integration tests


## Checklist
- [ ] I have added or updated any relevant documentation.
- [x] I have cleaned any remaining cloud resources from my accounts.