fix: only data node ignores lock if restarting (#762)
## Description of issue or feature:
If there is only one data node and is restarting it will be stuck
waiting for the lock because after we stop opensearch we cannot get the
lock from the index.

## Solution:
<!--- Please describe in detail what the solution implemented in this PR
is. The changes made etc. -->
This PR introduces an `ignore_lock` parameter when emitting the
`_start_opensearch_event`, allowing the lock to be bypassed if the
current unit is the only data app and is the leader.

OpenSearch restart logic improvements:

* Added logic to determine if the current unit is the only data node and
the leader, and set an `ignore_lock` flag accordingly when restarting
OpenSearch. This flag is passed to `_start_opensearch_event.emit()` to
allow the restart to proceed without waiting for a lock in single-node
scenarios.

## How was this change tested?
- [x] Manually
- [ ] Unit tests
- [ ] Integration tests