fix: remove verify repository from requirer error checks (#772)
## Description of issue or feature:

OpensearchPeerClusterRequirer performs some checks regarding
peer-cluster-relation. There was a check for repository verification but
this is not useful. This was aiming not publish creds to subclusters if
main orhestrator do not finish checks and preparations. However,
`self.refresh_relation_data` method is called in several places and
triggers peer_relation_changed event in the subclusters. By this way,
they may get the creds any time. This change does not prevent it, so
removing.


## Solution:

We are removing the verify_repository check from the method
`_error_set_from_requirer` which is not useful.

## How was this change tested?
- [x] Manually
- [ ] Unit tests
- [x] Integration tests

<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->


## Checklist
- [ ] I have added or updated any relevant documentation.
- [ ] I have cleaned any remaining cloud resources from my accounts.

Signed-off-by: gatici <gulsum.atici@canonical.com>