fix: Refactor code and verify repository after all units save creds (#774)
## Description of issue or feature:
<!--- Please describe in detail what this PR is about. And a reference
link to a Github issue or Jira ticket. -->
Fixes #771 
If the leader of the main orchestrator is different from the elected CM
in OS the verification of the repository will fail. We need to propagate
the credentials to all units of OS to make sure the elected CM has the
credentials before we actually verify that the credentials are valid.
## Solution:
<!--- Please describe in detail what the solution implemented in this PR
is. The changes made etc. -->
We propagate credentials to all units. Whenver a unit saves the
credentials it addes its hash into the unit data bag. Once all units
have credentials, the leader of the main orchestrator will verify the
credentials.

This PR also includes refactoring code to use a single event handler for
both s3/azure backends.
## How was this change tested?
- [x] Manually
- [ ] Unit tests
- [x] Integration tests